cmd/ab: implement qualcomm ab slot logic
Some Qualcomm systems come with a slotted boot logic implemented between
the system and the firmware to handle failed OTA upgrade. This is very
similar to Android and ChromiumOS, with a slight twist on how the logic
is handled.

Provide a command able to identify the next slot to be booted by reading
the GPT entries GUID specific attributes and update them in consequence.
The process is similar to ab_select for Android except it relies on GPT
metadata rather than a specific partition ("misc").

Following the example of ab_select, the cmd itself sets an environment
with the slot to be booted. The bootloader then has to load the right
boot slot containing the Android Boot Image and update the kernel
command like in the boot script to notify the OS of which slot it should
use for the other partitions.

The code is organized like ab_select, with the slot logic being under
boot/. The Qualcomm Android Boot Image partition type UUID is also added
to part_efi.h to be used by qcom_ab.c.

Signed-off-by: Eric Chanudet <echanude@redhat.com>