cmd/ab: implement chromeos ab slot logic
ChromiumOS provides and document a slotted boot logic implemented
between the system and the firmware to recover from failed OTA upgrade.

See:
https://chromium.googlesource.com/chromiumos/docs/+/HEAD/disk_format.md#selecting-the-kernel

Provide a command able to identify the next slot to be booted by reading
the GPT entries GUID specific attributes and update them in consequence.
The process is similar to ab_select for Android except it relies on GPT
metadata rather than a specific partition ("misc").

Signed-off-by: Eric Chanudet <echanude@redhat.com>